TestTestgresCommon - unified tests for local and remote
Let's stop creating duplicates for local and remote variants.

In this commit we moved the following tests:
 - test_version_management
 - test_child_pids
 - test_exceptions

run_tests.sh is updated to execute new non-remote tests from TestTestgresCommon, too.
